placement/tests: fix format-truncation warning in grp_str snprintf

+++ src/placement/tests/pl_debug.c
@@ -491,7 +491,7 @@ cmd_gen_layout(const char *arg)
 						val);
 					return;
 				}
-				snprintf(grp_str, sizeof(grp_str), "%ld", v);
+				snprintf(grp_str, sizeof(grp_str), "%u", (unsigned int)v);
 			}
 		} else {
 			fprintf(stderr, "Unknown gen_layout option: '%s'\n", tok);
